+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.7.7 (unreleased)
 
   * Fix #344 to avoid: "warning: default `to_a' will be obsolete"
+  * Fix #356 by adding 'known object classes' like struct to be recognized as objects.
 
 ## 0.7.6